23-11-2020

+++ database/migrations/2020_11_21_091252_create_pr_consultancy_cost_direct_table.php
@@ -17,7 +17,7 @@ class CreatePrConsultancyCostDirectTable extends Migration
            $table->engine = 'InnoDB';
             $table->id();
             $table->bigInteger('pr_consultancy_cost_id')->unsigned();
-            $table->integer('direct_cost')->unsigned();
+            $table->bigInteger('direct_cost')->unsigned();
             $table->timestamps();
             $table->foreign('pr_consultancy_cost_id')->references('id')->on('pr_consultancy_costs')->onDelete('cascade');
         });
